+++ app/about/page.tsx
@@ -1,30 +1,18 @@
 import Image from 'next/image'
 import OfficerCardTray from '@/components/about/OfficerCardTray';
+import AboutUsInfoSection from '@/components/about/AboutUsInfoSection';
 import InfoCardTray from '@/components/about/InfoCardTray';
 import { satoshiBold, satoshiMedium } from '@/util/localFonts';
 
 export default function About() {
 
-    // update this content/information with better branding.
-    const aboutUsInformation: string[][] = [
-      ["About Us", "We are a student organization inspired by AI, Machine Learning, and Data Science.", "text-4xl md:text-5xl lg:text-6xl"],
-      ["What we do.","CougarAI aims to educate members in the latest advances in artificial intelligence, data science, and machine learning. As well as to provide a space for networking among those interested in the field.", "text-3xl md:max-2xl:text-5xl"],
-    ];
-
     return (
       <main className="bg-cai-400 min-h-screen flex flex-col lg:grid lg:grid-cols-2">
         <section className="md:ml-24 lg:ml-32 lg:mt-2 md:max-lg:w-3/5">
           {/* the padding top (for lg screens and up) here might need a tweak to be consistent with the padding elsewhere on the page */}
-          <div className={satoshiBold.className + " flex flex-col gap-y-10 md:gap-y-16 lg:gap-y-24 text-white text-center md:text-start px-5 md:px-0 md:items-start pt-20 md:pt-64 lg:pt-36 mb-3.5"}>
-            {aboutUsInformation.map(([title, description, textSize], index) => (
-              <div key = {index}>
-                <h1 className={`pb-2 md:pb-5 align-bottom ${textSize}`}>{title}</h1>
-                <p className={satoshiMedium.className + " pb-2 md:pb-6 align-bottom text-base md:text-3xl leading-normal"}>{description}</p>
-              </div>
-            ))}
-          </div>
 
-          <InfoCardTray></InfoCardTray>
+          <AboutUsInfoSection/>
+          <InfoCardTray/>
 
           {/* Officer section of the page */}
           <div className= {satoshiBold.className + " flex flex-col text-center md:text-start md:items-start items-center my-10 lg:mt-28 mb-16 mr-64 md:pb-16 w-screen"}>
